Pull the global state and helpers into a separate file.
